minor re-arranging
